Splitting FormatterDocument into own top level class to avoid LineDocument imports in TextDocumentServiceImpl
